requirements.txt
customtkinter
Pillow
PyPDF2
pdf2image
pytesseract